move undeform_structure function to core utils
