+++ emmet-core/emmet/core/utils.py
@@ -7,6 +7,7 @@ from typing import Any, Dict, Iterator, List, Optional, Union
 import numpy as np
 from monty.json import MSONable
 from pydantic import BaseModel
+from pymatgen.analysis.elasticity.strain import Deformation
 from pymatgen.analysis.graphs import MoleculeGraph
 from pymatgen.analysis.local_env import OpenBabelNN, metal_edge_extender
 from pymatgen.analysis.molecule_matcher import MoleculeMatcher
@@ -16,6 +17,9 @@ from pymatgen.analysis.structure_matcher import (
     StructureMatcher,
 )
 from pymatgen.core.structure import Molecule, Structure
+from pymatgen.transformations.standard_transformations import (
+    DeformStructureTransformation,
+)
 from pymatgen.util.graph_hashing import weisfeiler_lehman_graph_hash
 
 from emmet.core.mpid import MPculeID
@@ -74,6 +78,32 @@ def group_structures(
     for _, pregroup in groupby(sorted(structures, key=_get_sg), key=_get_sg):
         for group in sm.group_structures(list(pregroup)):
             yield group
+
+
+def undeform_structure(structure: Structure, transformations: Dict) -> Structure:
+    """
+    Get an undeformed structure by applying transformations in a reverse order.
+
+    Args:
+        structure: deformed structure
+        transformation: transformation that deforms the structure
+
+    Returns:
+        undeformed structure
+    """
+
+    for transformation in reversed(transformations.get("history", [])):
+        if transformation["@class"] == "DeformStructureTransformation":
+            deform = Deformation(transformation["deformation"])
+            dst = DeformStructureTransformation(deform.inv)
+            structure = dst.apply_transformation(structure)
+        else:
+            raise RuntimeError(
+                "Expect transformation to be `DeformStructureTransformation`; "
+                f"got {transformation['@class']}"
+            )
+
+    return structure
 
 
 def group_molecules(molecules: List[Molecule]):
